nouvelle-backend/backend/requirements.txt
asgiref==3.11.1
Django==6.0.4
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
django-cors-headers==4.7.0
psycopg==3.3.3
psycopg-binary==3.3.3
PyJWT==2.12.1
python-dotenv==1.2.2
requests==2.32.3
sqlparse==0.5.5
tzdata==2026.1
reportlab==4.4.10
pillow==12.2.0
